Split out search related models and actions from main ts files
