+++ packages/components/src/components/search/SearchResultCard.spec.tsx
@@ -15,7 +15,7 @@
  */
 import * as React from 'react'
 import renderer from 'react-test-renderer'
-import { Map, fromJS } from 'immutable'
+import { fromJS } from 'immutable'
 import { SearchResultCard } from "./SearchResultCard";
 import { mount } from "enzyme";
